move FakePublisher to a module that is only loaded when needed
